added cerbot installation to documentation; updated command coverage table with notes to use docker; added a print to terminal option in start; updates to documentation;

+++ pocketlab/commands/start.py
@@ -14,9 +14,8 @@ _start_details = {
 }
 
 from pocketlab.init import fields_model
-from colorama import Fore, Style
 
-def start(service_list, verbose=True, virtualbox='default', environment_type='dev', ip_address=''):
+def start(service_list, verbose=True, virtualbox='default', environment_type='dev', ip_address='', print_terminal=''):
 
     title = 'start'
 
@@ -29,7 +28,8 @@ def start(service_list, verbose=True, virtualbox='default', environment_type='de
         'verbose': verbose,
         'virtualbox': virtualbox,
         'environment_type': environment_type,
-        'ip_address': ip_address
+        'ip_address': ip_address,
+        'print_terminal': print_terminal
     }
     for key, value in input_fields.items():
         if value:
@@ -124,7 +124,8 @@ def start(service_list, verbose=True, virtualbox='default', environment_type='de
             if container_alias == alias['NAMES'].split()[0]:
                 container_exists = True
         if container_exists:
-            raise ValueError('A container already exists for alias "%s".\nTry first: "docker rm -f %s"' % (container_alias, container_alias))
+            if not print_terminal:
+                raise ValueError('A container already exists for alias "%s".\nTry first: "docker rm -f %s"' % (container_alias, container_alias))
         details['alias'] = container_alias
         
     # validate mount paths exist
@@ -134,8 +135,7 @@ def start(service_list, verbose=True, virtualbox='default', environment_type='de
                 if volume['type'] == 'bind':
                     volume_path = path.join(details['path'], volume['source'])
                     if not path.exists(volume_path):
-                        if verbose:
-                            print('%sWARNING: Value "%s" for field volume[%s].source in %s is not a valid path.\nVolume "%s" will not be mounted.%s' % (Fore.MAGENTA, volume['source'], str(i), compose_insert, volume['source'], Style.RESET_ALL))
+                        raise ValueError('Value "%s" for field volume[%s].source in %s is not a valid path.\nVolume "%s" will not be mounted.' % (volume['source'], str(i), compose_insert, volume['source']))
 
     # validate ports are available
         if 'ports' in details['config'].keys():
@@ -226,8 +226,7 @@ def start(service_list, verbose=True, virtualbox='default', environment_type='de
             for volume in service_config['volumes']:
                 if volume['type'] == 'bind':
                     volume_path = path.join(service_path, volume['source'])
-                    if path.exists(volume_path):
-                        run_kwargs['mounted_volumes'][volume_path] = volume['target']
+                    run_kwargs['mounted_volumes'][volume_path] = volume['target']
         if 'command' in service_config.keys():
             run_kwargs['start_command'] = service_config['command']
         if 'networks' in service_config.keys():
@@ -235,7 +234,33 @@ def start(service_list, verbose=True, virtualbox='default', environment_type='de
                 run_kwargs['network_name'] = service_config['networks'][0]
 
     # run docker run
-        docker_client.run(**run_kwargs)
+        if print_terminal:
+
+        # compose run command
+            windows_path = ''
+            from platform import uname
+            local_os = uname()
+            if local_os.system in ('Windows'):
+                windows_path = '/'
+            sys_cmd = 'docker run --name %s' % run_kwargs['container_alias']
+            for key, value in run_kwargs['environmental_variables'].items():
+                sys_cmd += ' -e %s=%s' % (key.upper(), value)
+            for key, value in run_kwargs['mapped_ports'].items():
+                sys_cmd += ' -p %s:%s' % (key, value)
+            for key, value in run_kwargs['mounted_volumes'].items():
+                sys_cmd += ' -v %s"${pwd}/%s":%s' % (windows_path, path.relpath(key), value)
+            if run_kwargs['network_name']:
+                sys_cmd += ' --network %s' % run_kwargs['network_name']
+            sys_cmd += ' -d %s' % run_kwargs['image_name']
+            if run_kwargs['image_tag']:
+                sys_cmd += ':%s' % run_kwargs['image_tag']
+            if run_kwargs['start_command']:
+                sys_cmd += ' %s' % run_kwargs['start_command'].strip()
+            print(sys_cmd)
+            return exit_msg
+        
+        else:
+            docker_client.run(**run_kwargs)
 
     # report outcome
         port_msg = ''
